perf(ui): keep startup bundles within enforced budgets
