+++ ui/src/components/app-sidebar.ts
@@ -22,7 +22,7 @@ import type { CatalogSessionKey } from "../lib/sessions/catalog-key.ts";
 import type { CatalogProjectGrouping } from "../lib/sessions/catalog-project-grouping.ts";
 import { showToast } from "../lib/toast.ts";
 import { SubscriptionsController } from "../lit/subscriptions-controller.ts";
-import { SETTINGS_SEARCH_TARGETS } from "../pages/config/settings-targets.ts";
+import { SETTINGS_ROUTE_TARGETS } from "../pages/config/route-data.ts";
 import type { NewSessionTarget } from "../pages/new-session/location.ts";
 import { sidebarPluginTabs } from "./app-sidebar-nav-menus.ts";
 import {
@@ -144,7 +144,15 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
     expandedRows: [],
     visibleRows: [],
   };
-  private readonly narrationSubscriptions = this.createNarrationSubscriptions();
+  private readonly subscriptions = new SubscriptionsController(this)
+    .effect(
+      () => this.context?.gateway,
+      (gateway) => gateway.subscribeEvents((event) => this.narration?.handleEvent(event)),
+    )
+    .watch(
+      () => this.context?.agentIdentity,
+      (agentIdentity, notify) => agentIdentity.subscribe(notify),
+    );
   private readonly nativeGatewaysChanged = () => this.sidebarMenus.closeSessionMenu();
   private readonly refreshAppearanceSettings = () => this.context?.theme.refresh();
   private readonly hiddenSessionCatalogsChanged = () => {
@@ -163,17 +171,11 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
       }
     },
   );
-  private readonly agentIdentitySubscriptions = new SubscriptionsController(this).watch(
-    () => this.context?.agentIdentity,
-    (agentIdentity, notify) => agentIdentity.subscribe(notify),
-  );
-
   @state() catalogProjectGrouping = loadStoredSidebarCatalogGrouping();
 
   constructor() {
     super();
-    void this.narrationSubscriptions;
-    void this.agentIdentitySubscriptions;
+    void this.subscriptions;
     void new BoardAvailabilityController(
       this,
       () => {
@@ -205,15 +207,6 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
         };
       },
     );
-  }
-
-  private createNarrationSubscriptions(): SubscriptionsController {
-    const subscriptions = new SubscriptionsController(this);
-    subscriptions.effect(
-      () => this.context?.gateway,
-      (gateway) => gateway.subscribeEvents((event) => this.narration?.handleEvent(event)),
-    );
-    return subscriptions;
   }
 
   override disconnectedCallback() {
@@ -445,7 +438,7 @@ class AppSidebar extends AppSidebarSessionNavigationElement implements SessionLi
     setStoredSessionCatalogHidden(catalogId, true);
     // Reuse the settings-search destination for the Sidebar preferences block so the
     // toast opens the same place the rest of the app calls "Appearance > Sidebar".
-    const recovery = SETTINGS_SEARCH_TARGETS.appearanceSidebar;
+    const recovery = SETTINGS_ROUTE_TARGETS.appearanceSidebar;
     const recoveryHref =
       pathForRoute(recovery.routeId, this.basePath) + recovery.search + recovery.hash;
     // The section disappears instantly and its only standing recovery lives on another
